feat(manage): 完善试卷功能
- 删除冗余的 ExamPaperQuestionVo 类
- 新增 PaperQuestionDto 类用于组卷
- 修改 ExamPaperController 中的 composePaper 方法，改为 generateQuestions
- 新增 composePaper 方法用于组卷
- 优化题目生成和组卷逻辑

+++ eduflex-manage/src/main/java/com/eduflex/manage/domain/ExamPaper.java
@@ -41,7 +41,7 @@ public class ExamPaper extends BaseEntity {
      * 总分
      */
     @Excel(name = "总分")
-    private Long totalScore;
+    private Integer totalScore;
 
     @Excel(name = "关联课程")
     private Long courseId;
